+++ CHANGELOG.md
@@ -8,6 +8,7 @@ BREAKING:
     * remove the `Makefile`
     * remove the `switchver` subcommand that is no longer needed now that terraform binary is directly used from `~/.terraform.d/versions` and symbolic links are no longer created in `.wrapper/bin`
     * add `-V`/`--version` parameter to get the version of the tfwrapper itself
+    * `use_local_azure_session_directory` now defaults to `True`
 
 # 7.13.2 (2021/06/24)